Bugfix: Properly detect a partial dirty tree.
If a "git commit $file" was run with a clean index we would not notice
that other files in the working tree may have been modified.  Hence the
backup stash would be dropped after the tidy instead of being
re-applied in the post-commit hook. This would result in lost
modifications!