Adds support for Perl::Tidy::Sweet
Perl::Tidy::Sweet uses Perl::Tidy's prefilter and postfilter hooks to
support syntax added by a number of modules (p5-mop,
Method::Signatures::Simple, MooseX::Method::Signatures, MooseX::Declare,
etc).  This includes method and func keywords, including the (possibly
multi-line) parameter lists.

This patch adds a --sweet/-s flag to both the install and pre-commit
commands. When specified, code is tied with Perl::Tidy::Sweet instead of
Perl::Tidy.

Perl::Tidy::Sweet is added as a prereq in the Makefile.PL, as
"recommends" doesn't seem to do anything (I'm not that familiar with
Module::Install).

A new test has been added that optionally runs all the tests with the
--sweet flag. This might be overkill. It would probably be sufficient to
just test the tidying of code with the method keyword.

This is my contribution to May's PRC. I just happen to be the author of
Perl::Tidy::Sweet.